# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Landa Travels conversational planning API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/chatbot/overview:
    get:
      operationId: getAdminChatbotOverview
      tags: [admin]
      summary: Get live chatbot telemetry overview
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: "#/components/parameters/AdminFrom" }
        - { $ref: "#/components/parameters/AdminTo" }
      responses:
        "200": { description: Overview, content: { application/json: { schema: { $ref: "#/components/schemas/AdminAnalyticsResponse" } } } }
        "401": { description: Unauthenticated }
        "403": { description: Non-admin }
  /admin/chatbot/funnel:
    get:
      operationId: getAdminChatbotFunnel
      tags: [admin]
      summary: Get chatbot funnel and source breakdown
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: "#/components/parameters/AdminFrom" }
        - { $ref: "#/components/parameters/AdminTo" }
      responses:
        "200": { description: Funnel, content: { application/json: { schema: { $ref: "#/components/schemas/AdminAnalyticsResponse" } } } }
  /admin/chatbot/failures:
    get:
      operationId: listAdminChatbotFailures
      tags: [admin]
      summary: List supported chatbot failure cohorts
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: "#/components/parameters/AdminFrom" }
        - { $ref: "#/components/parameters/AdminTo" }
        - { $ref: "#/components/parameters/AdminFailureType" }
        - { $ref: "#/components/parameters/AdminSource" }
        - { $ref: "#/components/parameters/AdminOutcome" }
      responses:
        "200": { description: Failure cohorts, content: { application/json: { schema: { $ref: "#/components/schemas/AdminAnalyticsResponse" } } } }
  /admin/chatbot/conversations:
    get:
      operationId: listAdminChatbotConversations
      tags: [admin]
      summary: List privacy-safe chatbot conversations
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: "#/components/parameters/AdminFrom" }
        - { $ref: "#/components/parameters/AdminTo" }
        - { $ref: "#/components/parameters/AdminSource" }
        - { $ref: "#/components/parameters/AdminOutcome" }
        - { $ref: "#/components/parameters/AdminConversationFailureClass" }
        - { name: q, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100 } }
        - { name: offset, in: query, schema: { type: integer, minimum: 0 } }
      responses:
        "200": { description: Conversations, content: { application/json: { schema: { $ref: "#/components/schemas/AdminAnalyticsResponse" } } } }
  /admin/chatbot/conversations/{sessionId}:
    get:
      operationId: getAdminChatbotConversation
      tags: [admin]
      summary: Get redacted chatbot transcript and events
      security: [{ bearerAuth: [] }]
      parameters: [{ name: sessionId, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: Transcript, content: { application/json: { schema: { $ref: "#/components/schemas/AdminAnalyticsResponse" } } } }
  /admin/chatbot/evaluations:
    get:
      operationId: listAdminChatbotEvaluations
      tags: [admin]
      summary: List labelled chatbot evaluation runs
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Evaluation runs, content: { application/json: { schema: { $ref: "#/components/schemas/AdminAnalyticsResponse" } } } }
  /admin/chatbot/evaluations/{id}:
    get:
      operationId: getAdminChatbotEvaluation
      tags: [admin]
      summary: Get labelled evaluation predictions
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      responses:
        "200": { description: Evaluation, content: { application/json: { schema: { $ref: "#/components/schemas/AdminAnalyticsResponse" } } } }
  /admin/chatbot/evaluations/{id}/confusion-matrix:
    get:
      operationId: getAdminChatbotConfusionMatrix
      tags: [admin]
      summary: Get a labelled evaluation confusion matrix
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      responses:
        "200": { description: Matrix, content: { application/json: { schema: { $ref: "#/components/schemas/AdminAnalyticsResponse" } } } }
  /admin/chatbot/evaluations/{id}/export:
    get:
      operationId: exportAdminChatbotEvaluation
      tags: [admin]
      summary: Export labelled evaluation predictions
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
        - { name: format, in: query, schema: { type: string, enum: [json, csv] } }
      responses:
        "200": { description: Export }
  /admin/chatbot/reviews:
    post:
      operationId: createAdminChatbotReview
      tags: [admin]
      summary: Add a human review label
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminReviewInput"
      responses:
        "201": { description: Review created }
components:
  parameters:
    AdminFrom:
      name: from
      in: query
      description: Inclusive UTC date (YYYY-MM-DD)
      schema: { type: string, format: date }
    AdminTo:
      name: to
      in: query
      description: Inclusive UTC date (YYYY-MM-DD)
      schema: { type: string, format: date }
    AdminFailureType:
      name: type
      in: query
      schema:
        type: string
        enum: [high-fallback, repeated-prompt, abandoned-form, source-destination-confusion]
    AdminSource:
      name: source
      in: query
      schema: { type: string, enum: [demo, live] }
    AdminOutcome:
      name: outcome
      in: query
      schema: { type: string }
    AdminConversationFailureClass:
      name: failureClass
      in: query
      schema:
        type: string
        enum: [fallback, unresolved_intent, error]
  schemas:
...
    AdminAnalyticsResponse:
      type: object
      additionalProperties: true
      description: Analytics response. Telemetry and labelled evaluation responses identify their data source explicitly.
    AdminReviewInput:
      type: object
      required: [sessionId, label]
      properties:
        eventId: { type: string }
        sessionId: { type: string }
        label: { type: string, maxLength: 80 }
        notes: { type: string, maxLength: 1000 }
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer